Configure custom GitLab base URL in client
Leveraged the SetBaseURL function from the go-gitlab library to enable configuration of the GitLab API client with a custom base URL. This adjustment allows Mergestat to interface seamlessly with self-hosted GitLab instances, facilitating interactions with the API of such deployments. By utilizing this supported functionality, we ensure compatibility and maintainability of the customization, making Mergestat more adaptable to varied user environments.